requirements.txt
streamlit
Pillow
streamlit-folium
geocoder
folium
streamlit-webrtc
